add support for hoa and start on generating code
a bit stuck on how to read hoa still in terms of input and output on the transitions
in the SPOT docs, it says The label a & b & c should be understood as: "if the input is a&b, the output should be c".

do we just need to parse a, b, and c in this case to figure out which is input and which is output.
seems like there should be a better way
if not though, this approach isnt all that bad.

this makes progress on issue #2